Refactor: [#90] WebConfig 클래스에 logFilter 메소드 생성
RequestLoggingFilterConfig 클래스에 있던 logFilter 메소드를 WebConfig 클래스로 이동
