--- BE/src/main/java/com/enolj/airbnb/config/RequestLoggingFilterConfig.java
@@ -1,20 +1,0 @@
-package com.enolj.airbnb.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.filter.CommonsRequestLoggingFilter;
-
-@Configuration
-public class RequestLoggingFilterConfig {
-
-    @Bean
-    public CommonsRequestLoggingFilter logFilter() {
-        CommonsRequestLoggingFilter filter = new CommonsRequestLoggingFilter();
-        filter.setIncludeHeaders(false);
-        filter.setIncludeQueryString(true);
-        filter.setIncludePayload(true);
-        filter.setMaxPayloadLength(10000);
-        filter.setAfterMessagePrefix("REQUEST DATA : ");
-        return filter;
-    }
-}
